Calling for an Exercise in Dr Hu's R Workshop

The function calls a run_tutorial function to open a learnr project. The project list is continually updated along with the R workshop with Dr. Hu.

drhur(workshop, lang = "cn")

Arguments

workshop

A character vector indicating the name of the workshop to open. See details for the workshop list.

lang

A character vector indicating the display language. The current version includes workshops in English (en) and simple Chinese (cn). The default is Chinese.

Details

The list of workshop available for the current version:

calculate: Doing basic maths with R syntax.

play: Basic data structure in R; inputting, outputting, and creating data.

wrangle: Generalizing (such as sorting and summarizing) and extracting (selecting or adjusting certain columns or rows) data frames.

describe: Skills to describe variable and data.

visualize: Describing variables, exploring patterns, and presenting estimates with visualization tools.