Calling for an Exercise in Dr Hu's R Workshop

The function calls a run_tutorial function to open a learnr project. The project list is continually updated along with the R workshop with Dr. Hu.

drhur(workshop)

Arguments

workshop
A character vector indicating the name of the workshop to open. See details for the workshop list.

Details

The list of workshop available so far:

Playing around with R (playing): Intro to basic data structure in R; inputting, outputting, and creating data.

Data munging (munging): Generalizing (such as sorting and summarizing) and extracting (selecting or adjusting certain columns or rows) data frames.

Data analyzing (analyzing): Analyze data with descriptive and multivariate tools and diagnoses.

Data presenting (visualizing): Present descriptive and analytic results with `ggplot2`